# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/service-category-vendor": {
            "get": {
                "tags": [
                    "ServiceCategoryVendor"
                ],
                "summary": "Barcha vendor-category ulanishlarini olish",
                "operationId": "fd3152a71ddc6eb7cf6fe9f31a0cd4c7",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "ServiceCategoryVendor"
                ],
                "summary": "Yangi vendor-category ulanishi yaratish",
                "operationId": "446be9a1b094e0447b4863c9572ae4d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/service-category-vendor/{id}": {
            "delete": {
                "tags": [
                    "ServiceCategoryVendor"
                ],
                "summary": "Ulanishni o‘chirish",
                "operationId": "232a191c8f31b35fd8cd528db3ab7641",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Barcha xizmatlarni olish",
                "operationId": "dd9fa49eb984799a15fc472bf6fb5969",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Yangi xizmat yaratish",
                "operationId": "c019636ce99faf73705213724085e455",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Xizmatni ko‘rish",
                "operationId": "993742f57a2eb52737802f8b077c0319",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Xizmatni yangilash",
                "operationId": "303f9bb5de76c85e4b25683e9a919cb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Xizmatni o‘chirish",
                "operationId": "2d04ad98e76301c58f721d97a6fd9b36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/service-images": {
            "get": {
                "tags": [
                    "ServiceImages"
                ],
                "summary": "Barcha xizmat rasmlarini olish",
                "operationId": "f125aea841e9d07514572367df429900",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "ServiceImages"
                ],
                "summary": "Yangi xizmat rasmi yaratish",
                "operationId": "e91d16883ddc38bc68da59b179ffbefd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/service-images/{id}": {
            "get": {
                "tags": [
                    "ServiceImages"
                ],
                "summary": "Xizmat rasmini ko‘rish",
                "operationId": "498e1dc17b939151822a2e0be5c8c947",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "ServiceImages"
                ],
                "summary": "Xizmat rasmini yangilash",
                "operationId": "61ca4f91b6277af44c47de879e89ce66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ServiceImages"
                ],
                "summary": "Xizmat rasmini o‘chirish",
                "operationId": "1757730e381d4cd9aae73f63a5edaf0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/vendors": {
            "get": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Barcha vendorlarni olish",
                "operationId": "bc5dff1655a032ec33872d0c1d6b5488",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Yangi vendor yaratish",
                "operationId": "59c1c72301a690ad44b3ecf229e14989",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/vendors/{id}": {
            "get": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Vendorni ko‘rish",
                "operationId": "0ea12bbb87e0de87de6542d0a9c346b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Vendorni yangilash",
                "operationId": "0a9a2db89ac3b5cd23193872e5d01223",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Vendorni o‘chirish",
                "operationId": "97e2cc862f79b666353adffa00717028",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        }
    },
...
}